feat: remove comments
