--- apps/miranum-bpmn-modeler-webview/src/app/createElementTemplate/InputBindingProvider.ts
@@ -1,26 +1,0 @@
-// import {
-//     createInputParameter,
-//     ensureExtension,
-//     shouldUpdate,
-// } from "./utils/CreateHelper";
-// import { getDefaultValue } from "./utils/Helper";
-//
-// export default class InputBindingProvider {
-//     static create(element: any, options: any) {
-//         const { property, bpmnFactory } = options;
-//
-//         const { binding } = property;
-//
-//         const value = getDefaultValue(property);
-//
-//         const ioMapping = ensureExtension(element, "camunda:inputOutput", bpmnFactory);
-//
-//         if (!shouldUpdate(value, property)) {
-//             return;
-//         }
-//
-//         const input = createInputParameter(binding, value, bpmnFactory);
-//         input.$parent = ioMapping;
-//         ioMapping.get("inputParameters").push(input);
-//     }
-// }
