--- apps/miranum-bpmn-modeler-webview/src/app/createElementTemplate/PropertyBindingProvider.ts
@@ -1,18 +1,0 @@
-// import { getBusinessObject } from "bpmn-js/lib/util/ModelUtil";
-// import { getDefaultValue } from "./utils/Helper";
-//
-// export default class PropertyBindingProvider {
-//     static create(element: any, options: any) {
-//         const { property } = options;
-//
-//         const { binding } = property;
-//
-//         const { name } = binding;
-//
-//         const value = getDefaultValue(property);
-//
-//         const businessObject = getBusinessObject(element);
-//
-//         businessObject[name] = value;
-//     }
-// }
