--- apps/miranum-bpmn-modeler-webview/src/app/createElementTemplate/index.js
@@ -1,1 +1,0 @@
-export * from "./TemplateElementFactory";
