+++ apps/miranum-bpmn-modeler-webview/src/app/index.ts
@@ -1,3 +1,4 @@
 export * from "./ContentController";
 export * from "./utils";
 export * from "./formKeys";
+export * from "./TemplateElementFactory";
